% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Coordinates of the borders of all of the countries in the World}
\format{A \code{SpatialPolygonsDataFrame} object of :
\itemize{
  \item data
  \item polygons
}}
\usage{
countries
}
\description{
A \code{SpatialPolygonsDataFrame}
}
\keyword{datasets}
