% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_all_qualities}
\alias{get_all_qualities}
\title{Get All Qualities}
\usage{
get_all_qualities(country, delay = 1)
}
\arguments{
\item{country}{\code{character}, the name of the country}
}
\value{
A \code{numeric vector} containing the qualities from all wines from
the specified country
}
\description{
This function retrieves a list of qualities for a specified
country, corresponding to the wines from the WineEnthusiast webpage
}
