% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_all_titles}
\alias{get_all_titles}
\title{Get All Titles}
\usage{
get_all_titles(country, delay = 1)
}
\arguments{
\item{country}{\code{character}, the name of the country}
}
\value{
A \code{character} containing the qualities from all wines from
the specified country
}
\description{
This function retrieves a list of titles for a specified
country, corresponding to the wines from the WineEnthusiast webpage
}
