% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_pages}
\alias{get_pages}
\title{Get Pages}
\usage{
get_pages(country)
}
\arguments{
\item{country}{\code{character}, the name of the country}
}
\value{
A \code{vector} containing a sequence ranging from 1 to the total
number of pages
}
\description{
This function retrieves the total number of pages from the
WineEnthusiast webpage (www.winemag.com) for the specified country
}
