% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_prices}
\alias{get_prices}
\title{Get Prices}
\usage{
get_prices(country, page, validate = TRUE)
}
\arguments{
\item{country}{\code{character}, the name of the country}

\item{page}{\code{integer}, the webpage from which to retrieve the prices}
}
\value{
A \code{numeric vector} containing the prices from the wines on the
specified webpage
}
\description{
This function retrieves a list of prices for a specified
country for a single webpage,
corresponding to the wines from the WineEnthusiast website
}
