% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_quality}
\alias{get_quality}
\title{Get Quality}
\usage{
get_quality(country, page, validate = TRUE)
}
\arguments{
\item{country}{\code{character}, the name of the country}

\item{page}{\code{integer}, the webpage from which to retrieve the qualities}
}
\value{
A \code{numeric vector} containing the qualities from the wines on the
specified webpage
}
\description{
This function retrieves a list of qualities for a specified
country for a single webpage,
corresponding to the wines from the WineEnthusiast website
}
