% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape.R
\name{get_title}
\alias{get_title}
\title{Get Title}
\usage{
get_title(country, page, validate = TRUE)
}
\arguments{
\item{country}{\code{character}, the name of the country}

\item{page}{\code{integer}, the webpage from which to retrieve the titles}
}
\value{
A \code{character} containing the titles from the wines on the
specified webpage
}
\description{
This function retrieves a list of titles for a specified
country for a single webpage,
corresponding to the wines from the WineEnthusiast website
}
