% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\name{get_wine}
\alias{get_wine}
\title{Retrieving the best wine in the World}
\usage{
get_wine(Data, Country = NULL, Price = NULL, Variety = NULL,
  Province = NULL, N = 10, Criteria = NULL)
}
\arguments{
\item{Data}{A \code{dataframe} This is the dataframe that is used to retrieve
the wines. It should be the data: winemag_data_130k_v2.}

\item{Country}{A \code{character} This is the name of the country/countries
of interest.}

\item{Price}{A \code{numeric vector} This is a vector containing the minimum
and the maximum values of the price.}

\item{Variety}{A \code{character} This is the name of the variety/varieties
of interest.}

\item{Province}{A \code{character} This is the name of the province/provinces
of interest.}

\item{N}{A \code{integer} This is the number of wines returned by the
function. This number should be a positive integer.}

\item{Criteria}{A \code{character} The variable used to rank the wines. It
can take the following values: "price", "ascending price", "quality",
"price & quality", "ascending price & quality" or "quality/price".}
}
\value{
A \code{tibble} displaying the top N wines according to the specified
criterion.
}
\description{
This function will allow you to retrieve from the wine review
data, the best wine
}
\examples{
data = data("winemag")
get_wine(data, Country = "Italy")
}
\author{
Vanessa Armistead

Bart Roes

Luisa Pricken

Ameni Rouatbi
}
