% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winemag}
\alias{winemag}
\title{Wine reviews retrieved from WineEnthusiast on the 22nd of November 2017}
\format{A \code{DataFrame} object of 14 variables:
\itemize{
  \item X
  \item country
  \item description
  \item designation
  \item points
  \item price
  \item province
  \item region_1
  \item region_2
  \item taster_name
  \item taster_twitter_handle
  \item title
  \item variety
  \item winery
}}
\usage{
winemag
}
\description{
A \code{DataFrame}
}
\keyword{datasets}
