% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo_complete}
\alias{runDemo_complete}
\title{Run Shiny app}
\usage{
runDemo_complete()
}
\description{
This function allows you to launch the Shiny app when the
package is loaded, making use of the supplied winemag dataset
}
