% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo_webscraped}
\alias{runDemo_webscraped}
\title{Run Updated Shiny app}
\usage{
runDemo_webscraped()
}
\description{
This function allows for launching the Shiny app when the
package is loaded, using data that is freshly webscraped from Winemag.com
}
